requirements.txt
appdirs==1.4.4
awsebcli==3.20.6
beautifulsoup4==4.12.2
blinker==1.6.2
botocore==1.29.99
bs4==0.0.1
cachelib==0.10.2
cement==2.8.2
certifi==2023.5.7
charset-normalizer==2.0.12
Choco==1.0.5
click==8.1.3
colorama==0.4.3
cssselect==1.2.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.6
docopt==0.6.2
et-xmlfile==1.1.0
fake-useragent==1.1.3
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
load-dotenv==0.1.0
lxml==4.9.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
npm==0.1.1
numpy==1.24.3
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.1
pandas==2.0.1
parse==1.19.0
pathspec==0.10.1
pipreqs==0.4.13
platformdirs==3.5.1
plotly==5.14.1
prettytable==3.7.0
pyee==8.2.2
pypiwin32==223
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.26.0
requests-html==0.10.0
semantic-version==2.8.5
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
tenacity==8.2.2
termcolor==1.1.0
touch==2020.12.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
values==2020.12.3
virtualenv==20.23.0
w3lib==2.1.1
wcwidth==0.1.9
websockets==10.4
Werkzeug==2.3.4
yahoo-fin==0.8.9.1
yarg==0.1.9
zipp==3.15.0
